app/settings.py
# app/settings.py
import os

from dotenv import load_dotenv

load_dotenv()  # Load .env variables


class Settings:
    # Switch to test database if TESTING=1
    TESTING: bool = os.getenv("TESTING", "0") == "1"

    DB_HOST: str = os.getenv("DB_HOST")
    DB_PORT: str = os.getenv("DB_PORT")
    DB_USER: str = os.getenv("DB_USER")
    DB_PASSWORD: str = os.getenv("DB_PASSWORD")
    DB_NAME: str = os.getenv("DB_NAME")

    # Pick correct database URL
    DATABASE_URL: str = os.getenv("TEST_DATABASE_URL") if TESTING else f"mysql+mysqlconnector://{DB_USER}:{DB_PASSWORD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"

    # Auth settings
    SECRET_KEY: str = os.getenv("SECRET_KEY")
    ALGORITHM: str = os.getenv("ALGORITHM", "HS256")
    ACCESS_TOKEN_EXPIRE_MINUTES: int = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES", 30))


settings = Settings()
